move surface manipulation functions to nilearn.surface
