+++ nilearn/__init__.py
@@ -58,4 +58,4 @@ CHECK_CACHE_VERSION = True
 # list all submodules available in nilearn and version
 __all__ = ['datasets', 'decoding', 'decomposition', 'connectome',
            'image', 'input_data', 'masking', 'mass_univariate', 'plotting',
-           'region', 'signal', '__version__']
+           'region', 'signal', 'surface', '__version__']
